refactor(battery-soh): move persist_analytics to the service module
battery_health_v2.py previously did not export persist_analytics
(only compute_full_analytics was exported), but analytics.py and
battery_health.py both imported persist_analytics from it. The backend
crashed at startup with ImportError.

Fix:
- Move persist_analytics + _extract_kwh helper into battery_health_v2.py
  (the service module — that's where they belong)
- Remove the duplicate definitions from battery_health.py
- battery_health.py still imports persist_analytics from the service

Same semantics, just living in the right module now. Backend restarts
cleanly and serves 401 (auth working) on both endpoints.

Verified by hitting both endpoints from inside the backend container
via urllib — both return 401 Unauthorized with the standard error
envelope, confirming the routes are wired and the auth middleware is
active.
