+++ backend/app/services/battery_health_v2.py
@@ -39,12 +39,13 @@ from sqlalchemy import select, func, and_, desc
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.models.telemetry import (
-    UserVehicle,
-    ChargingSession,
-    ChargingCurve,
     BatteryHealth,
+    ChargingCurve,
+    ChargingSession,
     DriveRangeEstimatedFull,
+    BatteryHealthAnalytics,
 )
+from app.models.vehicle import UserVehicle
 
 
 # ===== Confidence levels =====
@@ -645,3 +646,54 @@ async def compute_full_analytics(
         anomalies=anomalies,
         computed_at=datetime.now(timezone.utc),
     )
+
+
+# ===== Persistence =====
+
+def _extract_kwh(m: MethodResult) -> float | None:
+    """Pick the closest-to-kWh numeric field from a method's extras for persistence.
+
+    Some methods don't produce a clean kWh number (e.g. cell_imbalance stores
+    imbalance_mv, throughput stores cycles). We persist whatever's closest to
+    the concept of "estimated capacity" so the v2 cache row isn't NULL.
+    """
+    if m.method == "tesla_capacity":
+        return m.extra.get("median_kwh")
+    if m.method == "cell_imbalance":
+        return m.extra.get("median_imbalance_mv")
+    if m.method == "throughput":
+        return m.extra.get("cycles")
+    if m.method == "range_drift_over_time":
+        return m.extra.get("retention_pct")
+    return None
+
+
+async def persist_analytics(db: AsyncSession, analytics: CombinedAnalytics) -> None:
+    """Write one row per method + one 'combined' row to battery_health_analytics."""
+    now = datetime.now(timezone.utc)
+    for m in analytics.methods:
+        row = BatteryHealthAnalytics(
+            user_vehicle_id=analytics.user_vehicle_id,
+            computed_at=now,
+            method=m.method,
+            soh_pct=m.soh_pct if m.soh_pct is not None else 0.0,
+            estimated_kwh=_extract_kwh(m),
+            sample_count=m.sample_count,
+            confidence=m.confidence,
+            inputs_json=m.inputs,
+            extra_json=m.extra,
+        )
+        db.add(row)
+    combined_row = BatteryHealthAnalytics(
+        user_vehicle_id=analytics.user_vehicle_id,
+        computed_at=now,
+        method="combined",
+        soh_pct=analytics.soh_pct if analytics.soh_pct is not None else 0.0,
+        estimated_kwh=analytics.estimated_kwh,
+        sample_count=sum(m.sample_count for m in analytics.methods),
+        confidence=analytics.confidence,
+        inputs_json=None,
+        extra_json={"anomalies": analytics.anomalies},
+    )
+    db.add(combined_row)
+    await db.commit()
